refactor: divide FileControl and Command into separate files
